requirement.txt
annotated-types==0.5.0
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
cryptography==41.0.3
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
pycparser==2.21
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
typing_extensions==4.7.1
Werkzeug==2.3.7
